+++ hydromt/raster.py
@@ -28,22 +28,13 @@ from scipy import ndimage
 import tempfile
 import pyproj
 import logging
+import rioxarray
 
 from . import gis_utils, HAS_PCRASTER
 
 logger = logging.getLogger(__name__)
 XDIMS = ("x", "longitude", "lon", "long")
 YDIMS = ("y", "latitude", "lat")
-FILL_VALUE_NAMES = ("_FillValue", "missing_value", "fill_value", "nodata", "nodatavals")
-UNWANTED_RIO_ATTRS = (
-    "nodatavals",
-    "crs",
-    "is_tiled",
-    "res",
-    "transform",
-    "scales",
-    "offsets",
-)
 GEO_MAP_COORD = "spatial_ref"
 
 
@@ -188,10 +179,7 @@ class XGeoBase(object):
     @property
     def crs(self):
         """Return Coordinate Reference System as :py:meth:`pyproj.CRS` object."""
-        if self.get_attrs("crs_wkt") is None:
-            self.set_crs()
-        crs = self.get_attrs("crs_wkt")
-        return CRS.from_wkt(crs) if crs is not None else crs
+        return self._obj.rio.crs
 
     @property
     def x_dim(self):
@@ -294,7 +282,7 @@ class XGeoBase(object):
                     except:
                         pass
         if input_crs is not None:
-            self.set_attrs(crs_wkt=input_crs.wkt)
+            self._obj.rio.write_crs(input_crs, inplace=True)
 
     def reset_spatial_dims_attrs(self):
         """Reset spatial dimension names and attributes to make CF-compliant
@@ -452,6 +440,45 @@ class XRasterBase(XGeoBase):
             and (np.isclose(dy, 0) or np.isclose(dy, 1))
             and np.logical_and.reduce((w <= w1, s <= s1, e >= e1, n >= n1))
         )
+
+    def gdal_compliant(self, rename_dims=True, force_sn=False):
+        """Updates attributes to get GDAL compliant NetCDF files.
+
+        Arguments
+        ----------
+        rename_dims: bool, optional
+            If True, rename x_dim and y_dim to standard names depending on the CRS
+            (x/y for projected and lat/lon for geographic).
+        force_sn: bool, optional
+            If True, forces the dataset to have South -> North orientation.
+
+        Returns
+        -------
+        ojb_out: xr.Dataset or xr.DataArray
+            GDAL compliant object
+        """
+        obj_out = self._obj
+        crs = obj_out.raster.crs
+        if (
+            obj_out.raster.res[1] < 0 and force_sn
+        ):  # write data with South -> North orientation
+            obj_out = obj_out.raster.flipud()
+        x_dim, y_dim, x_attrs, y_attrs = gis_utils.axes_attrs(crs)
+        if rename_dims:
+            obj_out = obj_out.rename(
+                {obj_out.raster.x_dim: x_dim, obj_out.raster.y_dim: y_dim}
+            )
+        else:
+            x_dim = obj_out.raster.x_dim
+            y_dim = obj_out.raster.y_dim
+        obj_out[x_dim].attrs.update(x_attrs)
+        obj_out[y_dim].attrs.update(y_attrs)
+        obj_out = obj_out.drop_vars(["spatial_ref"], errors="ignore")
+        obj_out.rio.write_crs(crs, inplace=True)
+        obj_out.rio.write_transform(obj_out.raster.transform, inplace=True)
+        obj_out.raster.set_spatial_dims()
+
+        return obj_out
 
     def transform_bounds(self, dst_crs, densify_pts=21):
         """Transform bounds from object to destination CRS.
@@ -1205,14 +1232,14 @@ class RasterDataArray(XRasterBase):
     @property
     def nodata(self):
         """Nodata value of the DataArray."""
-        _attrs = self._obj.attrs
-        _encoding = self._obj.encoding
-        if _attrs.get("_FillValue", None) is None:
-            self.set_nodata()
         # first check attrs, then encoding
-        return _attrs.get("_FillValue", _encoding.get("_FillValue", None))
+        nodata = self._obj.rio.nodata
+        if nodata is None:
+            nodata = self._obj.rio.encoded_nodata
+            self.set_nodata(nodata)
+        return nodata
 
-    def set_nodata(self, nodata=None):
+    def set_nodata(self, nodata=None, logger=logger):
         """Set the nodata value as CF compliant attribute of the DataArray.
 
         Arguments
@@ -1222,23 +1249,17 @@ class RasterDataArray(XRasterBase):
             If the nodata property and argument are both None, the _FillValue
             attribute will be removed.
         """
-        dtype = str(self._obj.dtype)
         if nodata is None:
-            for name in FILL_VALUE_NAMES:
-                if name in self._obj.attrs:
-                    # remove from attrs and set property
-                    nodata = self._obj.attrs.pop(name)
-                    if isinstance(nodata, tuple):
-                        nodata = nodata[0]
-                    try:
-                        nodata = getattr(np, dtype)(nodata)
-                        break
-                    except ValueError:
-                        nodata = None
-        if nodata is not None:  # keep attribute property
-            self._obj.attrs.update({"_FillValue": nodata})
-        elif "_FillValue" in self._obj.attrs:
-            self._obj.attrs.pop("_FillValue")
+            nodata = self._obj.rio.nodata
+            if nodata is None:
+                nodata = self._obj.rio.encoded_nodata
+        # Only numerical nodata values are supported
+        if np.issubdtype(type(nodata), np.number):
+            self._obj.rio.set_nodata(nodata, inplace=True)
+            self._obj.rio.write_nodata(nodata, inplace=True)
+        else:
+            logger.warning("No numerical nodata value found, skipping set_nodata")
+            self._obj.attrs.pop("_FillValue", None)
 
     def mask_nodata(self, fill_value=np.nan):
         """Mask nodata values with fill_value (default np.nan).
@@ -1257,7 +1278,7 @@ class RasterDataArray(XRasterBase):
         if self.nodata is not None:
             da_masked = self._obj.where(mask != 0, self.nodata)
         else:
-            logger.warn("Nodata value missing, skipping mask")
+            logger.warning("Nodata value missing, skipping mask")
             da_masked = self._obj
         return da_masked